pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cabbage-search"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" }, # TODO: Fill in your details
]
description = "An open-source Python library for AI-powered web search and summarization."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", # TODO: Choose and add a license file (e.g., LICENSE.txt)
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: General",
]
# Core dependencies identified from imports in the 'cabbage' package
dependencies = [
    "trafilatura>=2.0.0",
    "nltk>=3.9.0",
    "aiohttp>=3.9.0",
    "sumy>=0.11.0",
    "python-dotenv>=1.0.0",
    "selenium>=4.8.0",
    "webdriver-manager>=4.0.0",
    "duckduckgo-search>=8.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/cabbage-search" # TODO: Update URL
"Bug Tracker" = "https://github.com/yourusername/cabbage-search/issues" # TODO: Update URL

# Optional dependencies for running examples or development
[project.optional-dependencies]
api = [
    "uvicorn>=0.30.0",
    "fastapi>=0.110.0",
]
examples = [
    "requests>=2.28.0",
]
dev = [
    "cabbage-search[api,examples]", # Installs core + api + examples extras
    "pytest", # Example testing framework
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cabbage*"] # Finds the 'cabbage' package
exclude = ["tests*"] # Exclude tests directory if you add one later